Unificacion de un unico jframe principal y eliminacion de modulos especificos
