+++ src/main/java/com/mycompany/flash_project/Main.java
@@ -7,11 +7,9 @@ package com.mycompany.flash_project;
 
 
 import conexion_y_funciones.*;
-import interfaz.JFrame_gerente;
-import interfaz.JFrame_operador;
 import interfaz.JFrame_inicio;
-import interfaz.JFrame_secretaria;
 import interfaz.Factura;
+import interfaz.JFrame_principal;
 import interfaz.Ventanafactura;
 import java.net.URISyntaxException;
 import java.sql.Connection;
@@ -48,11 +46,11 @@ public class Main {
      * @param args the command line arguments
      */
     public static void main(String[] args) throws SQLException, URISyntaxException {
-        new JFrame_gerente().setVisible(true);
+        new JFrame_principal().setVisible(true);
 
 
-        new JFrame_operador().setVisible(true);
-        new JFrame_secretaria().setVisible(true);
+        //new JFrame_operador().setVisible(true);
+        //new JFrame_secretaria().setVisible(true);
         
         
         SessionFactory factory = HibernateUtil.buildSessionFactory();
